requirements.txt
# Discord dependencies
discord.py>=2.3.2
aiohttp>=3.9.1

# Environment variables
python-dotenv==1.0.0

# Data handling and parsing
pydantic>=2.5.2
cachetools>=5.3.2
pytz>=2023.3.post1
python-dateutil>=2.8.2
feedparser>=6.0.10

# Excel file handling
openpyxl>=3.1.2

# Utilities
typing-extensions>=4.8.0

# Time and date handling
tzdata>=2023.3